move gecko tests
